fix to make backtest module test work

+++ service/backtest/backtest.go
@@ -23,7 +23,7 @@ type Order struct {
 
 type Position struct {
 	Symbol        string    `json:"symbol" mapstructure:"symbol"`
-	Amount        int       `json:"amount" mapstructure:"amount"`
+	Amount        int64     `json:"amount" mapstructure:"amount"`
 	CostBasis     float64   `json:"cost_basis" mapstructure:"cost_basis"`
 	LastSalePrice float64   `json:"last_price" mapstructure:"last_sale_price"`
 	LastSaleDate  time.Time `json:"last_sale_date" mapstructure:"last_sale_date"`
